requirements.txt
mem0ai==0.1.39
google-generativeai==0.8.3
chromadb==0.4.24
python-dotenv==1.0.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-api-python-client==2.108.0
